requirements.txt
pandas
yfinance
python-dateutil
openpyxl
matplotlib
seaborn